requirements.txt
flask==3.0.0
flask_cors==4.0.0
python-dotenv
yagmail
gunicorn
gevent